asdf

+++ Ejercicio/src/Component.java
@@ -24,6 +24,6 @@ public interface Component {
 	*/
     float getPrecio();
     int getCodigo();
-    void setPrecio(int precio);
-    
+    void setPrecio(Float precio);
+
 }
